requirements.txt
aws-cdk-lib==2.54.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.51.1a0
aws-cdk.aws_apigatewayv2_alpha==2.51.1a0
aws-cdk.aws-lambda-python-alpha==2.51.1a0
aws_cdk.aws_apigatewayv2_alpha==2.56.0a0
